Fix DD-1211 bagId not correctly minted
Dataverse seems to use the latest published version to inherit existing
metadata fields into the draft version. The issue is that a
deaccessioned version would be used to compare to, and the bag ID in the
draft would be different from the deaccesioned version, thus no new bag
ID was minted.

This commit fixes that by comparing it to all previous versions that are
either published or deaccesioned to make sure no duplicate bag ID ends
up in the drafts metadata.